requirements.txt
Package                       Version
----------------------------- ----------
appdirs                       1.4.4
asgiref                       3.2.10
astroid                       2.4.2
autopep8                      1.5.4
black                         21.7b0
boto3                         1.14.31
botocore                      1.17.31
cachelib                      0.1
certifi                       2020.4.5.1
chardet                       3.0.4
click                         7.1.2
colorama                      0.4.4
cs50                          5.0.4
distlib                       0.3.1
Django                        3.1.4
django-ckeditor               5.9.0
django-cors-headers           3.5.0
django-crispy-forms           1.9.2
django-filter                 2.3.0
django-js-asset               1.2.2
django-storages               1.9.1
djangorestframework           3.11.1
djangorestframework-simplejwt 4.4.0
docutils                      0.15.2
filelock                      3.0.12
Flask                         1.1.2
Flask-Session                 0.3.2
gunicorn                      20.0.4
idna                          2.9
install                       1.3.3
isort                         5.6.4
itsdangerous                  1.1.0
Jinja2                        2.11.2
jmespath                      0.10.0
lazy-object-proxy             1.4.3
MarkupSafe                    1.1.1
mccabe                        0.6.1
mypy-extensions               0.4.3
pathspec                      0.9.0
Pillow                        8.0.1
pip                           21.2.1
psycopg2                      2.8.5
pycodestyle                   2.6.0
PyJWT                         1.7.1
pylint                        2.6.0
python-dateutil               2.8.1
python-dotenv                 0.13.0
pytz                          2020.1
regex                         2021.7.6
requests                      2.23.0
s3transfer                    0.3.3
setuptools                    57.0.0
six                           1.15.0
SQLAlchemy                    1.3.17
sqlparse                      0.3.1
termcolor                     1.1.0
toml                          0.10.2
tomli                         1.1.0
urllib3                       1.25.9
virtualenv                    20.2.1
Werkzeug                      0.15.4
wheel                         0.36.2
whitenoise                    5.1.0
wrapt                         1.12.1
